Commit: Informatics - 2024-2025 - Practicums - Group 3 - Week 02

+++ Informatics/2024-2025/Practicums/Group 3/Week 02/Solutions/Problem 02.c
@@ -3,14 +3,14 @@
 
 int main() {
     int     a   = 0     ;
-    double  b   = 0.0   ;
+    float   b   = 0.0   ;
     char    c   = '\0'  ;
 
-    scanf("%d %lf %c", &a, &b, &c);
+    scanf("%d %f %c", &a, &b, &c);
 
-    printf("%d\n",    a);
-    printf("%lf\n",   b);
-    printf("%c\n",    c);
+    printf("%d\n",  a);
+    printf("%f\n",  b);
+    printf("%c\n",  c);
 
     return 0;
 }
